# docs/05-api/openapi.yaml
openapi: 3.1.0
info:
  title: Rabit API
  version: 1.0.0-mvp
  description: |
    Contract-first MVP API (ADR-0010). Rules in api-guidelines.md.
    Operations marked `x-rabit-status: planned` are reserved and not implemented in MVP.
servers:
  - url: http://localhost:8080
    description: Local api (control plane)
  - url: http://localhost:8081
    description: Local media gateway (data plane)
security:
  - bearer: []
...
paths:
...
  /v1/recordings/{recording_id}/passport:
    parameters: [ { name: recording_id, in: path, required: true, schema: { $ref: '#/components/schemas/Id' } } ]
    get:
      tags: [integrity]
      operationId: getPassport
      responses:
        '200': { description: Music Passport, content: { application/json: { schema: { $ref: '#/components/schemas/Passport' } } } }
        '404': { $ref: '#/components/responses/Error' }
...
  /v1/reports:
    post:
      tags: [integrity]
      operationId: createReport
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/CreateReportRequest' } } }
      responses:
        '201': { description: Report received, content: { application/json: { schema: { $ref: '#/components/schemas/Report' } } } }
        '404': { $ref: '#/components/responses/Error' }
        '429': { $ref: '#/components/responses/Error' }
...
  /v1/ops/entitlements:
    post:
      tags: [ops]
      operationId: opsGrantEntitlement
      parameters: [ { $ref: '#/components/parameters/IdempotencyKeyRequired' } ]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/GrantEntitlementRequest' } } }
      responses:
        '201':
          description: Granted
          headers: { ETag: { $ref: '#/components/headers/ETag' } }
          content: { application/json: { schema: { $ref: '#/components/schemas/Entitlement' } } }
        '403': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
  /v1/ops/entitlements/{entitlement_id}/status:
    parameters: [ { name: entitlement_id, in: path, required: true, schema: { $ref: '#/components/schemas/Id' } } ]
    post:
      tags: [ops]
      operationId: opsChangeEntitlementStatus
      parameters: [ { $ref: '#/components/parameters/IfMatch' } ]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/ChangeStatusRequest' } } }
      responses:
        '200':
          description: Changed
          headers: { ETag: { $ref: '#/components/headers/ETag' } }
          content: { application/json: { schema: { $ref: '#/components/schemas/Entitlement' } } }
        '403': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
        '412': { $ref: '#/components/responses/Error' }
        '428': { $ref: '#/components/responses/Error' }
  /v1/ops/users/{user_id}/subscription:
    parameters: [ { name: user_id, in: path, required: true, schema: { $ref: '#/components/schemas/Id' } } ]
    put:
      tags: [ops]
      operationId: opsSetSubscription
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/SetSubscriptionRequest' } } }
      responses:
        '200': { description: Subscription state set (operator/sandbox only), content: { application/json: { schema: { $ref: '#/components/schemas/SubscriptionEnvelope' } } } }
        '403': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
...
  /v1/ops/reports:
    get:
      tags: [ops]
      operationId: opsListReports
      parameters:
        - { name: status, in: query, schema: { type: string, enum: [received, triaged, actioned, dismissed] } }
        - { $ref: '#/components/parameters/Limit' }
      responses:
        '200':
          description: Reports for triage (reporter identity omitted)
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      type: object
                      required: [report_id, subject_type, subject_id, reason_code, details, status, created_at]
                      properties:
                        report_id: { $ref: '#/components/schemas/Id' }
                        subject_type: { type: string }
                        subject_id: { $ref: '#/components/schemas/Id' }
                        reason_code: { type: string }
                        details: { type: [string, 'null'] }
                        status: { type: string }
                        created_at: { $ref: '#/components/schemas/Timestamp' }
        '403': { $ref: '#/components/responses/Error' }
  /v1/ops/reports/{report_id}/status:
    parameters: [ { name: report_id, in: path, required: true, schema: { $ref: '#/components/schemas/Id' } } ]
    post:
      tags: [ops]
      operationId: opsChangeReportStatus
      requestBody:
        required: true
        content: { application/json: { schema: { type: object, additionalProperties: false, required: [status, reason], properties: { status: { type: string, enum: [triaged, actioned, dismissed] }, reason: { $ref: '#/components/schemas/Reason' } } } } }
      responses:
        '200': { description: Changed, content: { application/json: { schema: { $ref: '#/components/schemas/Report' } } } }
        '403': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
  /v1/ops/integrity-signals:
    post:
      tags: [ops]
      operationId: opsRecordIntegritySignal
      description: Records a reviewed decision on one integrity axis. Only reviewed `recommendation_eligibility = excluded` removes an entity from DIG candidates.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [subject_entity_id, axis, value, reason]
              properties:
                subject_entity_id: { $ref: '#/components/schemas/Id' }
                axis: { type: string, enum: [ai_generation, technical_quality, spam_risk, rights_status, recommendation_eligibility] }
                value: { type: string, minLength: 1, maxLength: 50 }
                reason: { $ref: '#/components/schemas/Reason' }
      responses:
        '201':
          description: Recorded
          content: { application/json: { schema: { type: object, required: [axis, value, basis], properties: { axis: { type: string }, value: { type: string }, basis: { type: string, enum: [reviewed] } } } } }
        '403': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
...
components:
  securitySchemes:
    bearer: { type: http, scheme: bearer, bearerFormat: JWT }
  headers:
    ETag: { description: Aggregate version, schema: { type: string } }
  parameters:
...
    IdempotencyKeyRequired: { name: Idempotency-Key, in: header, required: true, schema: { $ref: '#/components/schemas/IdempotencyKey' } }
    IfMatch: { name: If-Match, in: header, required: false, description: Required for mutation (428 if missing), schema: { type: string } }
    Limit: { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
...
  responses:
    Error:
      description: Error envelope
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
  schemas:
    Id: { type: string, pattern: '^[a-z]{3}_[0-9A-HJKMNP-TV-Z]{26}$' }
    IdempotencyKey: { type: string, pattern: '^[A-Za-z0-9_-]{8,128}$' }
    Timestamp: { type: string, format: date-time }
...
    Reason: { type: string, minLength: 3, maxLength: 500 }
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, request_id, retryable]
          properties:
            code: { type: string }
            message: { type: string }
            request_id: { type: string }
            retryable: { type: boolean }
            details: { type: object }
...
    ArtistRef:
      type: object
      required: [entity_id, name]
      properties:
        entity_id: { $ref: '#/components/schemas/Id' }
        name: { type: string }
...
    Passport:
      type: object
      required: [recording_id, creation, credits, claims, integrity, report_url]
      properties:
        recording_id: { $ref: '#/components/schemas/Id' }
        creation:
          type: array
          description: Creation method per stage; stages without a claim are `unknown`.
          items:
            type: object
            required: [stage, method, verification_state, basis]
            properties:
              stage: { type: string, enum: [composition, lyrics, vocals, instruments, mixing, mastering, artwork] }
              method: { type: string, enum: [human, ai_assisted, ai_generated, unknown] }
              verification_state: { type: [string, 'null'], enum: [null, self_declared, distributor_verified, signature_valid, process_evidence_reviewed, rights_reviewed] }
              basis: { type: [string, 'null'], enum: [null, verified_fact, declared, ml_inferred] }
        credits: { type: array, items: { $ref: '#/components/schemas/Credit' } }
        claims:
          type: array
          items:
            type: object
            required: [claim_type, value, issuer, verification_state, basis]
            properties:
              claim_type: { type: string }
              value: { type: string }
              issuer: { type: string }
              verification_state: { type: string }
              basis: { type: string }
        integrity:
          type: array
          description: Independent axes; never a single quality score; internal scores are never exposed.
          items:
            type: object
            required: [axis, value, basis]
            properties:
              axis: { type: string, enum: [ai_generation, technical_quality, spam_risk, rights_status, recommendation_eligibility] }
              value: { type: string }
              basis: { type: string, enum: [declared, automated, reviewed] }
        report_url: { type: string }
    Credit:
      type: object
      required: [credit_id, contributor, role, instrument, creation_method, basis, verification_state, source]
      properties:
        credit_id: { $ref: '#/components/schemas/Id' }
        contributor: { $ref: '#/components/schemas/ArtistRef' }
        role: { type: string }
        instrument: { type: [string, 'null'] }
        creation_method: { type: string, enum: [human, ai_assisted, ai_generated, unknown] }
        basis: { type: string, enum: [verified_fact, declared, ml_inferred] }
        verification_state: { type: string }
        source: { type: string }
...
    Entitlement:
      type: object
      required: [entitlement_id, scope, resource_id, capabilities, origin, status, valid_from, valid_to, version]
      properties:
        entitlement_id: { $ref: '#/components/schemas/Id' }
        scope: { type: string, enum: [catalog_all, release, recording] }
        resource_id: { oneOf: [ { $ref: '#/components/schemas/Id' }, { type: 'null' } ] }
        capabilities: { type: array, items: { type: string } }
        origin: { type: string, enum: [subscription, purchase, grant] }
        status: { type: string, enum: [active, suspended, revoked, expired] }
        valid_from: { $ref: '#/components/schemas/Timestamp' }
        valid_to: { oneOf: [ { $ref: '#/components/schemas/Timestamp' }, { type: 'null' } ] }
        version: { type: integer }
...
    Subscription:
      type: object
      required: [subscription_id, plan, state, paid_through, source]
      properties:
        subscription_id: { $ref: '#/components/schemas/Id' }
        plan: { type: string }
        state: { type: string, enum: [active, past_due, cancelled, expired] }
        paid_through: { $ref: '#/components/schemas/Timestamp' }
        source: { type: string, enum: [operator, sandbox] }
    SubscriptionEnvelope:
      type: object
      required: [subscription]
      properties:
        subscription: { oneOf: [ { $ref: '#/components/schemas/Subscription' }, { type: 'null' } ] }
...
    CreateReportRequest:
      type: object
      additionalProperties: false
      required: [subject_type, subject_id, reason_code]
      properties:
        subject_type: { type: string, enum: [recording, release, artist, relation, credit] }
        subject_id: { $ref: '#/components/schemas/Id' }
        reason_code: { type: string, enum: [wrong_credit, wrong_relation, undisclosed_ai, spam, impersonation, rights_infringement, other] }
        details: { type: string, maxLength: 2000 }
    Report:
      type: object
      required: [report_id, status, created_at]
      properties:
        report_id: { $ref: '#/components/schemas/Id' }
        status: { type: string, enum: [received, triaged, actioned, dismissed] }
        created_at: { $ref: '#/components/schemas/Timestamp' }
...
    ChangeStatusRequest:
      type: object
      additionalProperties: false
      required: [status, reason]
      properties:
        status: { type: string, enum: [active, suspended, revoked] }
        reason: { $ref: '#/components/schemas/Reason' }
    GrantEntitlementRequest:
      type: object
      additionalProperties: false
      required: [user_id, scope, capabilities, reason]
      properties:
        user_id: { $ref: '#/components/schemas/Id' }
        scope: { type: string, enum: [release, recording] }
        resource_id: { $ref: '#/components/schemas/Id' }
        capabilities: { type: array, minItems: 1, items: { type: string, enum: [play] } }
        valid_to: { $ref: '#/components/schemas/Timestamp' }
        reason: { $ref: '#/components/schemas/Reason' }
    SetSubscriptionRequest:
      type: object
      additionalProperties: false
      required: [state, paid_through, reason]
      properties:
        plan: { type: string, default: listen_sandbox }
        state: { type: string, enum: [active, past_due, cancelled, expired] }
        paid_through: { $ref: '#/components/schemas/Timestamp' }
        reason: { $ref: '#/components/schemas/Reason' }
...